Quantum Key Distribution/backend/requirements.txt
# Runtime dependencies (Python 3.13 compatible)
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.6.1
Flask-CORS==6.0.2
Flask-Migrate==4.1.0
PyMySQL==1.1.2
cryptography==46.0.7
python-socketio==5.16.1
python-engineio==4.13.1
bcrypt==5.0.0
Werkzeug==3.1.8
python-dotenv==1.2.2
numpy==2.4.4
scipy==1.17.1
requests==2.33.1

# Development
pytest==8.4.2
pytest-flask==1.3.0
black==25.9.0
isort==7.0.0
flake8==7.3.0
